Enterprise installer should use C99 mode to compile.
This resolves issues with some platforms (e.g Cygwin) where specifying -ansi as the standard to compile against results in single-line (double slash) comments not being recognized as valid and producing compilation errors. This fixes this by upgrading to a newer standard which defines single-line comments so that obedient compilers will not produce errors.