Change configuration file to be named enterprise.cfg
Using a name of .MLUL-Live-USB is not really a good idea because it makes the file difficult to create and edit, especially for novice users. Also, Mac Linux USB Loader is a wholly separate tool from Enterprise; having its initials in Enterprise's configuration file is does convey the separation well. Using a different name will make everything much easier to understand.
Enterprise still supports .MLUL-Live-USB files, but issues a warning if it has to read it. If both files are found on the same drive, enterprise.cfg will "win" and its settings will be the ones which are used.