Let the user type in their own custom kernel parameters
If the user has to type in their own custom parameters to pass
to the Linux kernel, then they can now input them directly from
within the boot configuration section of Enterprise. This allows
users to not have to edit the configuration file if they want
to test loading Linux with an option that isn't hard-coded.