Fix memory allocation issue in keyboard input function
The maximum length of string that we allow the user to type is
1024 characters, but we were only allowing them to type 1023
before due to a coding error: I had incorrectly assumed that 1024
was the total length of string, but in fact it is not as that
figure doesn't include the null terminator.

Also, we were incorrectly handling the allocation of the initial
memory. Due to a failure to specify the order of operations, not
enough memory was being allocated for the null terminator.

Both these issues were one-line changes that are now fixed.